refactor(chat): extract model-availability errors to own module
Move the user-facing model-availability error responses (_raise_byok_key_required,
_raise_model_removed) and the stale-preference scrubber (_cleanup_stale_model_preferences,
_MODEL_PREF_KEYS) out of llm_config.py into model_availability.py, pulling the file
from 1006 to 917 lines (back under the 1k threshold). These are a distinct concern —
HTTP error shaping plus pref-row cleanup that resolve_llm_config merely invokes — not
part of the resolution engine, and no test mocks them, so the split is behavior- and
test-preserving. Re-imported into llm_config for its callers. 3546 unit tests pass.
